FEATURE: Add excluded items list (#1)
* FEATURE: Add excluded items list

This commit adds the ability to exclude certain items from the compact nav, having them appear as normal, on the side of the drop down menu.

* FIX: Add checks for mobile

This commit will add a check for mobile. If mobile, the excluded list will not be relevant, and a regular drop down list will be shown.